fix: add routes

+++ database/migrations/000002_table_statistics.up.sql
@@ -5,12 +5,12 @@ CREATE TABLE BranchDailyStats (
     date DATE NOT NULL,
     total_boletos INT DEFAULT 0,
     total_pagos INT DEFAULT 0,
-    valor_emitido NUMERIC(20, 2) DEFAULT 0.00,
-    valor_recebido NUMERIC(20, 2) DEFAULT 0.00,
+    valor_emitido DOUBLE PRECISION DEFAULT 0.00,
+    valor_recebido DOUBLE PRECISION DEFAULT 0.00,
     boletos_cancelados INT DEFAULT 0,
-    valor_cancelado NUMERIC(20, 2) DEFAULT 0.00,
+    valor_cancelado DOUBLE PRECISION DEFAULT 0.00,
     boletos_atrasados INT DEFAULT 0,
-    total_dias_atraso NUMERIC(20, 2) DEFAULT 0.00,
+    total_dias_atraso DOUBLE PRECISION DEFAULT 0.00,
     CONSTRAINT unique_tenant_branch_date UNIQUE (tenant_id, branch_id, date)
 );
 
@@ -21,9 +21,10 @@ CREATE TABLE CustomerMonthlyStats (
     month DATE NOT NULL,
     total_boletos INT DEFAULT 0,
     total_pagos INT DEFAULT 0,
-    valor_emitido NUMERIC(20, 2) DEFAULT 0.00,
-    valor_recebido NUMERIC(20, 2) DEFAULT 0.00,
+    valor_emitido DOUBLE PRECISION DEFAULT 0.00,
+    valor_recebido DOUBLE PRECISION DEFAULT 0.00,
     boletos_atrasados INT DEFAULT 0,
-    total_dias_atraso NUMERIC(20, 2) DEFAULT 0.00,
+    total_dias_atraso DOUBLE PRECISION DEFAULT 0.00,
     CONSTRAINT unique_customer_month UNIQUE (customer_document_number, month)
 );
+
